backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.24.1
requests==2.31.0
psycopg2-binary==2.9.9
PyJWT==2.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.2.0
gspread==6.0.0
google-auth==2.29.0
schedule==1.2.2
anthropic>=0.25.0
supabase==2.3.0
email-validator==2.2.0
